freebnb/freebnb/server/requirements.txt
asgiref==3.2.10
Django==3.1
django-localflavor==3.0.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
Pillow==7.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-stdnum==1.14
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
channels==2.4.0